refactor(tnt): share padding and signing between transient and targeted messages
